requirements.txt
playwright==1.39.0
requests>=2.31.0
loguru>=0.7.2
python-dotenv>=1.0.0
